ckeditor move folder #58
