+++ frontend/plugins/ckeditor.js
@@ -1,3 +1,3 @@
 import Vue from 'vue';
-import Ckeditor from '../components/Atom/Ckeditor.vue';
+import Ckeditor from '../components/Atom/Ckeditor/Ckeditor.vue';
 Vue.component('Ckeditor', Ckeditor);
